service interfaces
